Moved Fig S6
